Save starting address to a pointer.
At first, the dest pointer has the address of the first byte.
At the end of the loops, it points to the address of the last byte,
that cointains either nul or the last character of the array *which,
is returned. Therefore it is advisable to save the first address that is pointed
to by dest in another pointer p and returning it.